:hammer: refactored streaming engine
