+++ go.mod
@@ -3,7 +3,7 @@ module github.com/0chain/zboxmobile
 go 1.13
 
 require (
-	github.com/0chain/gosdk v0.0.0
+	github.com/0chain/gosdk v1.2.87
 	github.com/btcsuite/btcd v0.22.0-beta // indirect
 	github.com/deckarep/golang-set v1.7.1 // indirect
 	github.com/h2non/filetype v1.1.1 // indirect
@@ -17,10 +17,7 @@ require (
 	go.uber.org/multierr v1.7.0 // indirect
 	go.uber.org/zap v1.17.0 // indirect
 	golang.org/x/crypto v0.0.0-20210616213533-5ff15b29337e // indirect
-	golang.org/x/net v0.0.0-20210405180319-a5a99cb37ef4 // indirect
-	golang.org/x/sys v0.0.0-20210616094352-59db8d763f22 // indirect
-	gotest.tools v2.2.0+incompatible
 )
 
 // temporary, for development
-replace github.com/0chain/gosdk => /Users/artem.bogomaz/Documents/GIT2/0chain/go/gosdk
+//replace github.com/0chain/gosdk => ../gosdk
