+++ zbox/allocation.go
@@ -2,15 +2,7 @@ package zbox
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
-	"os"
-	"path"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
 
 	"github.com/0chain/gosdk/zboxcore/blockchain"
 	"github.com/0chain/gosdk/zboxcore/fileref"
@@ -362,8 +354,8 @@ func (a *Allocation) MoveObject(path string, destPath string) error {
 
 // GetMinWriteRead - getting back cost for allocation
 func (a *Allocation) GetMinWriteRead() (string, error) {
-	minW, minR, err := a.sdkAllocation.GetMinWriteRead()
-	maxW, maxR, err := a.sdkAllocation.GetMaxWriteRead()
+	minW, minR, _ := a.sdkAllocation.GetMinWriteRead()
+	maxW, maxR, _ := a.sdkAllocation.GetMaxWriteRead()
 
 	minMaxCost := &MinMaxCost{}
 	minMaxCost.maxR = maxR
@@ -391,402 +383,6 @@ func (a *Allocation) GetMinStorageCost(size int64) (string, error) {
 	return fmt.Sprintf("%f", float64(cost)), err
 }
 
-func (a *Allocation) GetFirstSegment(localPath, remotePath, tmpPath string, delay int) (string, error) {
-	if len(remotePath) == 0 {
-		return "", errors.New("Error: remotepath / authticket flag is missing")
-	}
-
-	if len(localPath) == 0 {
-		return "", errors.New("Error: localpath is missing")
-	}
-
-	dir := path.Dir(localPath)
-
-	file, err := os.Create(localPath)
-
-	if err != nil {
-		return "", err
-	}
-
-	downloader := &M3u8Downloader{
-		localDir:      dir,
-		localPath:     localPath,
-		remotePath:    remotePath,
-		authTicket:    "",
-		allocationID:  a.ID,
-		rxPay:         false,
-		downloadQueue: make(chan MediaItem, 100),
-		playlist:      sdk.NewMediaPlaylist(delay, file),
-		done:          make(chan error, 1),
-	}
-
-	downloader.allocationObj = a.sdkAllocation
-
-	listResult, err := a.sdkAllocation.ListDir(downloader.remotePath)
-	list := listResult.Children
-	if err != nil {
-		return "", err
-	}
-
-	downloader.Lock()
-	n := len(downloader.items)
-	max := 3 //len(list)
-	latestItem := ""
-
-	sort.Slice(list, func(i, j int) bool {
-		val_1, _ := strconv.Atoi(strings.ReplaceAll(strings.ReplaceAll(list[i].Name, "sync", ""), ".ts", ""))
-		val_2, _ := strconv.Atoi(strings.ReplaceAll(strings.ReplaceAll(list[j].Name, "sync", ""), ".ts", ""))
-		return val_1 < val_2
-	})
-
-	if n < max {
-		//sort.Sort(SortedListResult(list))
-
-		for i := n; i < max; i++ {
-			item := MediaItem{
-				Name: list[i].Name,
-				Path: list[i].Path,
-			}
-			downloader.items = append(downloader.items, item)
-			downloader.playlist.Append(item.Name)
-			downloader.playlist.Wait = append(downloader.playlist.Wait, item.Name)
-
-			downloader.localDir = tmpPath
-
-			if _, err = downloader.download(item); err == nil {
-				fmt.Println(err)
-			}
-			latestItem = item.Name
-			downloader.addToDownload(item)
-		}
-	}
-	downloader.Unlock()
-
-	downloader.playlist.Writer.Truncate(0)
-	downloader.playlist.Writer.Seek(0, 0)
-	downloader.playlist.Writer.Write(downloader.playlist.Encode())
-	downloader.playlist.Writer.Sync()
-
-	return latestItem, nil
-}
-
-// GetMinStorageCost - getting back min cost for allocation
-func (a *Allocation) PlayStreaming(localPath, remotePath, authTicket, lookupHash, initSegment string, delay int, statusCb StatusCallback) error {
-	downloader, err := createM3u8Downloader(localPath, remotePath, authTicket, a.ID, lookupHash, initSegment, false, delay)
-	if err != nil {
-		return err
-	}
-
-	downloader.status = statusCb
-	go downloader.Start()
-	return nil
-}
-
-// M3u8Downloader download files from blobber's dir, and build them into a local m3u8 playlist
-type M3u8Downloader struct {
-	sync.RWMutex
-	delay int
-
-	localDir     string
-	localPath    string
-	remotePath   string
-	authTicket   string
-	allocationID string
-	rxPay        bool
-
-	allocationObj *sdk.Allocation
-
-	lookupHash    string
-	items         []MediaItem
-	downloadQueue chan MediaItem
-	playlist      *sdk.MediaPlaylist
-	done          chan error
-	status        StatusCallback
-	initSegment   string
-}
-
-func createM3u8Downloader(localPath, remotePath, authTicket, allocationID, lookupHash, initSegment string, rxPay bool, delay int) (*M3u8Downloader, error) {
-	if len(remotePath) == 0 && (len(authTicket) == 0) {
-		return nil, errors.New("Error: remotepath / authticket flag is missing")
-	}
-
-	if len(localPath) == 0 {
-		return nil, errors.New("Error: localpath is missing")
-	}
-	dir := path.Dir(localPath)
-
-	file, err := os.Create(localPath)
-
-	if err != nil {
-		return nil, err
-	}
-
-	downloader := &M3u8Downloader{
-		localDir:      dir,
-		localPath:     localPath,
-		remotePath:    remotePath,
-		authTicket:    authTicket,
-		allocationID:  allocationID,
-		rxPay:         rxPay,
-		downloadQueue: make(chan MediaItem, 5),
-		playlist:      sdk.NewMediaPlaylist(delay, file),
-		done:          make(chan error, 1),
-		initSegment:   initSegment,
-	}
-
-	if len(remotePath) > 0 {
-		if len(allocationID) == 0 { // check if the flag "path" is set
-			return nil, errors.New("Error: allocation flag is missing") // If not, we'll let the user know
-		}
-
-		allocationObj, err := sdk.GetAllocation(allocationID)
-
-		if err != nil {
-			return nil, fmt.Errorf("Error fetching the allocation: %s", err)
-		}
-
-		downloader.allocationObj = allocationObj
-
-	} else if len(authTicket) > 0 {
-		allocationObj, err := sdk.GetAllocationFromAuthTicket(authTicket)
-		if err != nil {
-			return nil, fmt.Errorf("Error fetching the allocation: %s", err)
-		}
-
-		downloader.allocationObj = allocationObj
-
-		at := sdk.InitAuthTicket(authTicket)
-		isDir, err := at.IsDir()
-		if isDir && len(lookupHash) == 0 {
-			lookupHash, err = at.GetLookupHash()
-			if err != nil {
-				return nil, fmt.Errorf("Error getting the lookuphash from authticket: %s", err)
-			}
-
-			downloader.lookupHash = lookupHash
-		}
-		if !isDir {
-			return nil, fmt.Errorf("Invalid operation. Auth ticket is not for a directory: %s", err)
-		}
-	}
-
-	return downloader, nil
-}
-
-// Start start to download ,and build playlist
-func (d *M3u8Downloader) Start() error {
-	if d.status != nil {
-		d.status.Started(d.allocationID, d.localPath, 0, 0)
-	}
-
-	go d.autoRefreshList()
-	go d.autoDownload()
-	go d.playlist.Play() // TODO call only if queue downloaded fully per transaction
-
-	err := <-d.done
-
-	return err
-}
-
-func (d *M3u8Downloader) autoRefreshList() {
-	for {
-		list, err := d.getList()
-		if err != nil {
-			continue
-		}
-
-		d.Lock()
-		n := len(d.items)
-		max := len(list)
-
-		// sort.Sort(SortedListResult(list))
-		sort.Slice(list, func(i, j int) bool {
-			val_1, _ := strconv.Atoi(strings.ReplaceAll(strings.ReplaceAll(list[i].Name, "sync", ""), ".ts", ""))
-			val_2, _ := strconv.Atoi(strings.ReplaceAll(strings.ReplaceAll(list[j].Name, "sync", ""), ".ts", ""))
-			return val_1 < val_2
-		})
-
-		initId := n
-		if len(d.initSegment) > 0 {
-			initId = sort.Search(len(list), func(i int) bool {
-				return list[i].Name < d.initSegment
-			})
-
-			if n == max || n > initId {
-				initId = n
-			}
-		}
-
-		if initId < max {
-			// sort.Sort(SortedListResult(list))
-
-			for i := initId; i < max; i++ {
-				item := MediaItem{
-					Name: list[i].Name,
-					Path: list[i].Path,
-				}
-
-				/*
-					if len(d.items) > 0 {
-						found := sort.Search(len(d.items), func(i int) bool {
-							return d.items[i].Name == item.Name
-						})
-
-						if found == len(d.items) { // not found
-							println("=======initId ignoring as found in items:" + item.Name)
-							continue
-						}
-					}*/
-
-				println("===============addToDownload=" + list[i].Name)
-
-				d.items = append(d.items, item)
-				d.addToDownload(item)
-			}
-		}
-		d.Unlock()
-
-		time.Sleep(1 * time.Second)
-	}
-}
-
-func (d *M3u8Downloader) autoDownload() {
-	for {
-		item := <-d.downloadQueue
-
-		// file already exists
-		localPath := d.localDir + string(os.PathSeparator) + item.Name
-		_, err := os.Stat(localPath)
-		if err == nil {
-			println("===============downloaded============" + localPath)
-			d.playlist.Append(item.Name)
-			continue
-		}
-
-		for i := 0; i < 3; i++ {
-			if path, err := d.download(item); err == nil {
-				d.playlist.Append(path)
-				if d.status != nil {
-					println("===============downloaded============" + localPath)
-					d.status.InProgress(d.allocationID, path, 1, len(d.items), nil)
-				}
-				break
-			}
-		}
-	}
-}
-
-func (d *M3u8Downloader) addToDownload(item MediaItem) {
-	d.downloadQueue <- item
-}
-
-func (d *M3u8Downloader) download(item MediaItem) (string, error) {
-	wg := &sync.WaitGroup{}
-	statusBar := &StatusBarMocked{wg: wg}
-	wg.Add(1)
-
-	localPath := d.localDir + string(os.PathSeparator) + item.Name
-	remotePath := item.Path
-
-	if len(d.remotePath) > 0 {
-		println("==========start downloading file========= " + localPath)
-		err := d.allocationObj.DownloadFile(localPath, remotePath, statusBar)
-		if err != nil {
-			return "", err
-		}
-		wg.Wait()
-	}
-
-	if !statusBar.success {
-		return "", statusBar.err
-	}
-
-	return localPath, nil
-}
-
-func (d *M3u8Downloader) getList() ([]*sdk.ListResult, error) {
-	//get list from remote allocations's path
-	if len(d.remotePath) > 0 {
-		ref, err := d.allocationObj.ListDir(d.remotePath)
-		if err != nil {
-			return nil, err
-		}
-		return ref.Children, nil
-	}
-
-	//get list from authticket
-	if len(d.authTicket) > 0 {
-		ref, err := d.allocationObj.ListDirFromAuthTicket(d.authTicket, d.lookupHash)
-		if err != nil {
-			return nil, err
-		}
-		return ref.Children, nil
-	}
-
-	return nil, fmt.Errorf("file not found")
-}
-
-// SortedListResult sort files order by name
-/*
-type SortedListResult []*sdk.ListResult
-
-func (a SortedListResult) Len() int {
-	return len(a)
-}
-func (a SortedListResult) Less(i, j int) bool {
-
-	l := a[i]
-	r := a[j]
-
-	if len(l.Name) < len(r.Name) {
-		return true
-	}
-
-	if len(l.Name) > len(r.Name) {
-		return false
-	}
-
-	return l.Name < r.Name
-}
-
-func (a SortedListResult) Swap(i, j int) {
-	a[i], a[j] = a[j], a[i]
-}*/
-
-// MediaItem ts file
-type MediaItem struct {
-	Name string
-	Path string
-}
-
-type StatusBarMocked struct {
-	wg      *sync.WaitGroup
-	success bool
-	err     error
-}
-
-func (s *StatusBarMocked) Started(allocationId, filePath string, op int, totalBytes int) {}
-
-func (s *StatusBarMocked) InProgress(allocationId, filePath string, op int, completedBytes int, data []byte) {
-	println("filePath=========InProgress")
-}
-
-func (s *StatusBarMocked) Error(allocationID string, filePath string, op int, err error) {
-	println("filePath=========Error")
-	s.success = false
-	s.err = err
-	s.wg.Done()
-}
-
-func (s *StatusBarMocked) Completed(allocationId, filePath string, filename string, mimetype string, size int, op int) {
-	s.success = true
-	s.wg.Done()
-}
-
-func (s *StatusBarMocked) CommitMetaCompleted(request, response string, err error) {}
-
-func (s *StatusBarMocked) RepairCompleted(filesRepaired int) {}
-
 // GetMaxStorageCostWithBlobbers - getting cost for listed blobbers
 func (a *Allocation) GetMaxStorageCostWithBlobbers(size int64, blobbersJson string) (string, error) {
 	var selBlobbers *[]*sdk.BlobberAllocation
@@ -797,4 +393,14 @@ func (a *Allocation) GetMaxStorageCostWithBlobbers(size int64, blobbersJson stri
 
 	cost, err := a.sdkAllocation.GetMaxStorageCostFromBlobbers(size, *selBlobbers)
 	return fmt.Sprintf("%f", cost), err
+}
+
+// GetFirstSegment - getting the amount of segments in maxSegments for very first playback
+func (a *Allocation) GetFirstSegment(localPath, remotePath, tmpPath string, delay, maxSegments int) (string, error) {
+	return CreateStreamingService(a).GetFirstSegment(localPath, remotePath, tmpPath, delay, maxSegments)
+}
+
+// GetMinStorageCost - getting back min cost for allocation
+func (a *Allocation) PlayStreaming(localPath, remotePath, authTicket, lookupHash, initSegment string, delay int, statusCb StatusCallback) error {
+	return CreateStreamingService(a).PlayStreaming(localPath, remotePath, authTicket, lookupHash, initSegment, delay, statusCb)
 }
