+++ zbox/allocation_test.go
@@ -2,8 +2,6 @@ package zbox
 
 import (
 	"sort"
-	"strconv"
-	"strings"
 	"testing"
 
 	"github.com/0chain/gosdk/zboxcore/sdk"
@@ -18,14 +16,13 @@ func TestListSorting(t *testing.T) {
 	list = append(list, sdk.ListResult{Name: "test1.ts"})
 	list = append(list, sdk.ListResult{Name: "test5.ts"})
 	list = append(list, sdk.ListResult{Name: "test6.ts"})
+	list = append(list, sdk.ListResult{Name: "test10.ts"})
 	list = append(list, sdk.ListResult{Name: "test15.ts"})
 	list = append(list, sdk.ListResult{Name: "test14.ts"})
 	list = append(list, sdk.ListResult{Name: "test42.ts"})
 
 	sort.Slice(list, func(i, j int) bool {
-		val_1, _ := strconv.Atoi(strings.ReplaceAll(strings.ReplaceAll(list[i].Name, "test", ""), ".ts", ""))
-		val_2, _ := strconv.Atoi(strings.ReplaceAll(strings.ReplaceAll(list[j].Name, "test", ""), ".ts", ""))
-		return val_1 < val_2
+		return GetNumber(list[i].Name) < GetNumber(list[j].Name)
 	})
 
 	initId := sort.Search(len(list), func(i int) bool {
