+++ zbox/util.go
@@ -4,6 +4,7 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"regexp"
 	"strconv"
 
 	"github.com/0chain/gosdk/zboxcore/client"
@@ -94,4 +95,14 @@ func Sign(hash string) (string, error) {
 // VerifySignature - verify message with signature
 func VerifySignature(signature string, msg string) (bool, error) {
 	return client.VerifySignature(signature, msg)
+}
+
+func GetNumber(value string) int {
+	re := regexp.MustCompile("[0-9]+")
+	submatchall := re.FindAllString(value, -1)
+	for _, element := range submatchall {
+		res, _ := strconv.Atoi(element)
+		return res
+	}
+	return -1
 }
